feat #218 問い合わせ関連のテストケース追加
